+++ src/models/commitMessage.ts
@@ -1,0 +1,6 @@
+import { CommitType } from "./commitType";
+
+export type CommitMessage = CommitType & {
+  commitMessage: string;
+  commitBody: string;
+};