+++ src/models/preferences.ts
@@ -1,0 +1,7 @@
+import { OnSelection } from "./onSelection";
+import { TypeMode } from "./typeMode";
+
+export type Preferences = {
+  typeMode: TypeMode;
+  onSelection: OnSelection;
+};