+++ src/models/typeMode.ts
@@ -1,0 +1,4 @@
+export enum TypeMode {
+  DEFAULT = "default",
+  GITMOJI = "gitmoji",
+}