+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-# Issue Parser Changelog
+# Commit Issue Parser Changelog
 
 ## [Initial Version] - {PR_MERGE_DATE}
+
+- Added the `Parse Issue Url` command! 🎉
+- Jira and GitHub issue URLs are supported
+- The entry can contain the url, the commit description and the commit body separated by `,`